GL Entry Sampling with GUI
This was created during my internship with a Big4 accounting firm. One of my tasks was to sample entries from huge ledgers, and it was a pain because oftentimes the files were so large that it would crash the computer when displaying it. This application selects desired columns and randomly samples the input amount without opening the General Ledger file, thus making it multiple times as efficient.

I wrote a GUI and exported this as an independent application so I could run this on other computers without python.